cleanup and refactor tests
