+++ test/functional/revision_policies.js
@@ -4,12 +4,9 @@
 /* global describe, it, before, beforeEach, after, afterEach */
 
 var assert = require('assert');
-var cass = require('cassandra-driver');
-var dbu = require('../lib/dbutils.js');
+var dbu = require('../../lib/dbutils.js');
 var fs = require('fs');
-var makeClient = require('../lib/index');
-var TimeUuid = cass.types.TimeUuid;
-var yaml = require('js-yaml');
+var router = require('../utils/test_router.js');
 
 var testSchema = {
     table: 'revPolicyLatestTest',
@@ -44,14 +41,7 @@ var testSchema = {
 describe('MVCC revision policies', function() {
     var db;
     before(function() {
-        return makeClient({
-            log: function(level, info) {
-                if (/^error|fatal/.test(level)) {
-                    console.log(level, info);
-                }
-            },
-            conf: yaml.safeLoad(fs.readFileSync(__dirname + '/test_router.conf.yaml'))
-        })
+        return router.setup()
         .then(function(newDb) {
             db = newDb;
         })
