+++ test/functional/schema_migrations.js
@@ -4,11 +4,11 @@
 /* global describe, it, before, beforeEach, after, afterEach */
 
 var assert = require('assert');
-var dbu = require('../lib/dbutils');
+var dbu = require('../../lib/dbutils');
 var extend = require('extend');
 var fs = require('fs');
-var makeClient = require('../lib/index');
-var router = require('./test_router.js');
+var makeClient = require('../../lib/index');
+var router = require('../utils/test_router.js');
 var yaml = require('js-yaml');
 
 var hash = dbu.makeSchemaHash;
@@ -52,8 +52,8 @@ describe('Schema migration', function() {
         })
         .then(function(response) {
             assert.ok(response, 'undefined response');
+            console.log(response);
             assert.deepEqual(response.status, 201);
-
             router.makeRouter();
         });
     });
