+++ test/utils/jshint.spec.js
